# api-spec-v3.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses/{id}/wishlist':
    post:
      tags:
        - 'U33 : Wishlist Business'
      summary: Add business to user wishlist
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Id of the business to wishlist
          required: true
          example: 3
      responses:
        '201':
          description: Business has been successfully added to the users wishlist
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/wishlist/{id}':
    delete:
      tags:
        - 'U33: Wishlist business'
        - 'Wishlist'
      summary: Delete a wishlist entity
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          description: Wishlist entity ID
          schema:
            type: integer
          required: true
          example: 5
      responses:
        '200':
          description: Wishlist item deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: "Forbidden: Returned when a user tries to delete a wishlist item that they are not the owner of AND the user is not a D/GAA"
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      tags:
        - 'U33: Wishlist business'
        - 'Wishlist'
      summary: Update a wishlist mute status
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          description: Wishlist entity ID
          schema:
            type: integer
          required: true
          example: 5
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                mutedStatus:
                  type: string
                  description: The new muted status to change the business wishlist item to
                  example: "MUTED"
      responses:
        '200':
          description: Wishlist item mute status updated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: "Forbidden: Returned when a user tries to delete a wishlist item that they are not the owner of AND the user is not a D/GAA"
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/users/{id}/wishlist':
    get:
      tags:
        - 'U33: Wishlist business'
        - 'Wishlist'
      summary: Get a user's wishlist
      parameters:
        - name: id
          in: path
          description: The id of the user for whom to get wishlist.
          required: true
          schema:
            type: integer
            example: 100
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a (possibly empty) list of wishlist items is returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wishlist'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
...
    Wishlist:
      type: object
      description: An entity represnting a users wishlisted business
      required:
        - id
        - userId
        - businessId
      properties:
        id:
          type: integer
          description: Unique id of the notification
          example: 5
        userId:
          type: integer
          description: Id relating to the user
        businessId:
          type: integer
          description: Id of the wishlisted business
        mutedStatus:
          type: object
          description: MutedStatus describing whether the notifications for the wishlisted business are muted
...